move `TodoList.tsx` of `solid-drizzle` to `components` dir
