+++ boilerplates/solid-drizzle/package.json
@@ -39,11 +39,11 @@
     }
   },
   "exports": {
+    "./components/TodoList": {
+      "types": "./dist/types/components/TodoList.d.ts"
+    },
     "./pages/todo/+data": {
       "types": "./dist/types/pages/todo/+data.d.ts"
-    },
-    "./pages/todo/TodoList": {
-      "types": "./dist/types/pages/todo/TodoList.d.ts"
     },
     "./pages/todo/+Page": {
       "types": "./dist/types/pages/todo/+Page.d.ts"
@@ -54,11 +54,11 @@
   },
   "typesVersions": {
     "*": {
+      "components/TodoList": [
+        "./dist/types/components/TodoList.d.ts"
+      ],
       "pages/todo/+data": [
         "./dist/types/pages/todo/+data.d.ts"
-      ],
-      "pages/todo/TodoList": [
-        "./dist/types/pages/todo/TodoList.d.ts"
       ],
       "pages/todo/+Page": [
         "./dist/types/pages/todo/+Page.d.ts"
